feat: add persistent showing for tooltip component
